pepper: Remove stat.h and enable mknod in standard stat.h
Defining our own stat.h overrides the standard one, meaning that we
need to define **everything** in stat.h, including the S_IF* flags,
etc.

Instead, the standard stat.h has mknod declared but guarded by a
macro that checks for certain operating systems. We add __pepper__,
which is a define that our toolchain sets, to the list of supported
systems